Feat/frontend/chat/invite pong (#227)
* [backend] Update invite user test to use correct error
message

* [frontend] Add Pong game invitation functionality

* [backend] invite-cancel-pong event

* [backend] rename PrivateUserEntity -> PublicUserEntity

* [backend] [need review] ws に紐づけるuser の情報を 公開しても良い情報だけにした。

* [backend] add test to invite-cancel-pong

* [backend] to avoid duplicate class name. fix test

* [backend] delete don't use file and environment variable

* [frontend] fix ip address hard coding with oauth

+++ backend/src/chat/chat.gateway.ts
@@ -93,7 +93,14 @@ export class ChatGateway {
   @SubscribeMessage('invite-cancel-pong')
   handleInviteCancelPong(@ConnectedSocket() client: Socket) {
     const inviteUser = this.chatService.getUser(client);
+    const invitee = this.chatService.getInvite(inviteUser.id);
+    if (!invitee) {
+      this.server.to(client.id).emit('error-pong', 'No pending invite found.');
+      return;
+    }
+    const inviteeWsId = this.chatService.getWsFromUserId(invitee)?.id;
     this.chatService.removeInvite(inviteUser.id);
+    this.server.to(inviteeWsId).emit('invite-cancel-pong', inviteUser);
   }
 
   @SubscribeMessage('approve-pong')
